feat: Implement cgroup management for task execution
- Introduced a new `cgroup` module to handle cgroup creation, configuration, and management.
- Added `CgroupConfig` struct for specifying resource limits (CPU, memory, and PID limits) for tasks.
- Enhanced the `Runtime` struct to support cgroup configuration during task execution.
- Implemented functions to create and clean up task-specific cgroups, as well as to read task statistics.
- Updated error handling in `FaberError` to include cgroup-related errors.
- Modified task execution logic to utilize cgroup management for resource tracking.

+++ crates/faber-runtime/src/result.rs
@@ -164,6 +164,6 @@ impl<'de> serde::Deserialize<'de> for TaskResult {
 #[derive(Debug, Clone, Default, Deserialize, Serialize)]
 pub struct TaskResultStats {
     pub memory_peak_bytes: u64,
-    pub cpu_usage_percent: f64,
+    pub cpu_usage_percent: u64,
     pub execution_time_ms: u64,
 }
